segformer error

+++ models/model_segformer3d.py
@@ -7,7 +7,7 @@ from typing import Tuple, List
 @torch.jit.script
 def cube_root(n: int) -> int:
     """Compute integer cube root efficiently."""
-    return round(n ** (1.0 / 3.0))
+    return int(round(n ** (1.0 / 3.0)))
 
 
 class PatchEmbedding(nn.Module):
